+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 * make `ImageStatistics.valid_percent` a value between 0 and 100 (instead of 0 and 1) (author @param-thakker, https://github.com/cogeotiff/rio-tiler/pull/400)
 
+* warn user when dataset has Nodata value and Mask or Alpha band (https://github.com/cogeotiff/rio-tiler/pull/402)
+
 ## 2.1.0 (2021-05-17)
 
 * add auto-rescaling in `ImageData.render` method to avoid error when datatype is not supported by the output driver (https://github.com/cogeotiff/rio-tiler/pull/391)